+++ src/modules/gatsby-source-url/createImgixUrlFieldConfig.ts
@@ -1,8 +1,8 @@
 import { pipe } from 'fp-ts/lib/pipeable';
 import * as T from 'fp-ts/lib/Task';
 import * as TE from 'fp-ts/lib/TaskEither';
-import { GraphQLFieldConfig } from 'graphql';
-import { ComposeFieldConfigAsObject } from 'graphql-compose';
+import { GraphQLFieldConfig } from 'gatsby/graphql';
+import { ObjectTypeComposerAsObjectDefinition } from 'graphql-compose';
 import ImgixClient from 'imgix-core-js';
 import * as R from 'ramda';
 import {
@@ -12,7 +12,7 @@ import {
 import { IImgixParams, ImgixUrlArgs } from '../../publicTypes';
 import {
   gatsbySourceImgixUrlFieldType,
-  ImgixUrlParamsInputType,
+  ImgixParamsInputType,
   unTransformParams,
 } from './graphqlTypes';
 
@@ -35,7 +35,7 @@ export const createImgixUrlFieldConfig = <TSource, TContext>({
   description: 'A plain imgix URL with the URL and params applied.',
   args: {
     imgixParams: {
-      type: ImgixUrlParamsInputType,
+      type: ImgixParamsInputType,
       defaultValue: {},
     },
   },
@@ -61,9 +61,8 @@ export const createImgixUrlFieldConfig = <TSource, TContext>({
 
 export const createImgixUrlSchemaFieldConfig = <TSource, TContext>(
   args: CreateImgixUrlFieldConfigArgs<TSource>,
-): ComposeFieldConfigAsObject<TSource, TContext, ImgixUrlArgs> =>
-  createImgixUrlFieldConfig(args) as ComposeFieldConfigAsObject<
-    TSource,
-    TContext,
-    ImgixUrlArgs
-  >;
+): ObjectTypeComposerAsObjectDefinition<TSource, TContext> =>
+  ({
+    ...createImgixUrlFieldConfig(args),
+    name: 'ImgixGatsbyUrl',
+  } as ObjectTypeComposerAsObjectDefinition<TSource, TContext>);