+++ src/modules/gatsby-source-url/createRootImgixImageType.ts
@@ -1,10 +1,11 @@
+// TODO(luis): remove this file, no longer being imported anywhere
 import { GatsbyCache } from 'gatsby';
 import {
   GraphQLFieldConfig,
   GraphQLNonNull,
   GraphQLObjectType,
   GraphQLString,
-} from 'graphql';
+} from 'gatsby/graphql';
 import type ImgixClient from 'imgix-core-js';
 import * as R from 'ramda';
 import { IImgixGatsbyRootArgs, IImgixParams } from '../../publicTypes';
@@ -19,8 +20,7 @@ export const createRootImgixImageType = (
   imgixClient: ImgixClient,
   cache: GatsbyCache,
   defaultParams: Partial<IImgixParams>,
-  // TODO: fix any
-): GraphQLFieldConfig<any, any, IImgixGatsbyRootArgs> => ({
+): GraphQLFieldConfig<IRootSource, {}, IImgixGatsbyRootArgs> => ({
   args: {
     url: {
       type: GraphQLNonNull(GraphQLString),
@@ -28,29 +28,29 @@ export const createRootImgixImageType = (
         'The path of the image to render. If using a Web Proxy Source, this must be a fully-qualified URL.',
     },
   },
-  type: new GraphQLObjectType<any, any, any>({
+  type: new GraphQLObjectType({
     name: 'ImgixImage',
     fields: {
-      url: createImgixUrlFieldConfig<IRootSource, unknown>({
+      url: createImgixUrlFieldConfig({
         imgixClient,
         resolveUrl: R.prop('rawUrl'),
         defaultParams,
       }),
-      fluid: createImgixFluidFieldConfig<IRootSource, unknown>({
+      fluid: createImgixFluidFieldConfig({
         imgixClient,
         resolveUrl: R.prop('rawUrl'),
         cache,
         defaultParams,
-      }),
+      }) as GraphQLFieldConfig<IRootSource, {}, any>,
       fixed: createImgixFixedFieldConfig<IRootSource, unknown>({
         imgixClient,
         resolveUrl: R.prop('rawUrl'),
         cache,
         defaultParams,
-      }),
+      }) as GraphQLFieldConfig<IRootSource, {}, any>,
     },
   }),
-  resolve(_: any, args: IImgixGatsbyRootArgs): IRootSource {
+  resolve(_: unknown, args: IImgixGatsbyRootArgs): IRootSource {
     return { rawUrl: args.url };
   },
 });